# lib/api-zod/src/generated/types/appUser.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Disaster information and situational reporting API
 * OpenAPI spec version: 0.1.0
 */
import type { AppUserStatus } from './appUserStatus';
import type { UserRole } from './userRole';

export interface AppUser {
  id: string;
  name: string;
  email: string;
  role: UserRole;
  status: AppUserStatus;
  createdAt: Date;
}