# lib/api-zod/src/generated/types/userRole.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Disaster information and situational reporting API
 * OpenAPI spec version: 0.1.0
 */

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  Administrator: 'Administrator',
  Coordinator: 'Coordinator',
  Viewer: 'Viewer',
} as const;