# lib/api-zod/src/generated/types/userRoleInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Disaster information and situational reporting API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface UserRoleInput {
  role: UserRole;
}